rename gen modules -> generator
