+++ crates/examples/src/blake/round/generator.rs
