+++ crates/examples/src/blake/scheduler/generator.rs
