+++ crates/examples/src/blake/xor_table/generator.rs
